Create master_interface_2_control_vrep.py
This is the first working version of the control of the UMIRobot using the master interface 2. The code needs cleaning up since it has a lot of unnecessary leftover stuff which I'll update as I work on it.
Connect your 6 pots to the Arduino Shield and it should be able to operate using task space control.